app用户订单数量使用分页插件统计

app用户订单数量使用分页插件统计

See merge request !1197

+++ okdeer-mall-order/src/main/java/com/okdeer/mall/order/service/TradeOrderService.java
@@ -1066,6 +1066,6 @@ public interface TradeOrderService {
 	 * @author chenzc
 	 * @date 2017年2月28日
 	 */
-	int countUserOrders(UserOrderParamBo paramBo);
+	long countUserOrders(UserOrderParamBo paramBo);
 	// end V2.1 add by chenzc 2017-2-28
 }
